Use uintptr_t for pointer values in the ExecutionEngine.
llvm-svn: 10425

+++ llvm/include/llvm/ExecutionEngine/GenericValue.h
@@ -19,7 +19,7 @@
 
 namespace llvm {
 
-typedef uint64_t PointerTy;
+typedef uintptr_t PointerTy;
 
 union GenericValue {
   bool            BoolVal;
